Rooms can now be gotten from the database!
Added @JsonCreator and @JsonProperty to Room and Details so that they can be constructed (by the ObjectMapper I think)

Removed the getName() method from Room, because serialization doesn't work properly if that method exists.

Removed the annotations on Reservable @JsonTypeInfo and @JsonSubTypes - they only create errors.

Added empty constructors to RoomResponse and BuildingResponse as ObjectMapper doesn't work without them.

Added a method to convert models using convertValue instead of readValue

In RoomEntryComponent the controller has to be connected to an FXML before I can take and use the fields.