Building and room are now connected (thanks to Dimitar)
How did this happen? We added an additional attribute in Reservable.java, called building
While adding a room, the user inputs a building number. The ReservableController.java gets that number as a PathVariable and then the ReservableService.java gets the Building from the database and adds it to the room object. Then the room adds itself to the Building's List<Reservable>