Deleted obsolete files and implemented better app navigation everywhere
Deleted the pages for viewing a specific building or a specific room, since they are not going to be implemented functionalities (we're prioritizing testing the code and time is short)
Removed mainScene, which was used for development purposes only
With the new app navigation it would be much easier to achieve higher code coverage. And it's also tidier and more concise.
There was also an unnecessary button in AddBuildings and AddRooms leading to itself which is removed.
The example function to showcase the use of the new API was removed as well.
Stylesheets for fxml files are now loaded from them, and not externally